+++ module-order-service/src/main/java/com/shop/orderservice/client/ProductClient.java
@@ -1,6 +1,6 @@
 package com.shop.orderservice.client;
 
-import com.shop.orderservice.dto.response.ProductResponse;
+import com.shop.orderservice.order.dto.response.ProductResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
