+++ module-order-service/src/main/java/com/shop/orderservice/order/controller/OrderController.java
@@ -1,10 +1,10 @@
-package com.shop.orderservice.controller;
+package com.shop.orderservice.order.controller;
 
 import com.shop.orderservice.common.response.ResponseDto;
-import com.shop.orderservice.dto.request.OrderCreateRequest;
-import com.shop.orderservice.dto.response.OrderResponse;
-import com.shop.orderservice.entity.Order;
-import com.shop.orderservice.service.OrderService;
+import com.shop.orderservice.order.dto.request.OrderCreateRequest;
+import com.shop.orderservice.order.dto.response.OrderResponse;
+import com.shop.orderservice.order.entity.Order;
+import com.shop.orderservice.order.service.OrderService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
