+++ module-order-service/src/main/java/com/shop/orderservice/order/repository/OrderRepository.java
@@ -1,6 +1,6 @@
-package com.shop.orderservice.repository;
+package com.shop.orderservice.order.repository;
 
-import com.shop.orderservice.entity.Order;
+import com.shop.orderservice.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
